Separate view & edit components for dtype UI packages (markdown, default)
