+++ packages/dtype-default-ui/src/components/DefaultEdit.vue
@@ -1,27 +1,21 @@
 <template>
   <v-flex xs12>
-    <p>
-      Edit:
-      {{parseContent(aliasData)}}
+    Edit: <p>
+      {{parsedContent}}
     </p>
   </v-flex>
 </template>
 
 <script>
+import {VFlex} from 'vuetify/lib';
+
 export default {
   name: 'DefaultEdit',
-  props: ['data'],
-  mounted() {
-    this.parseContent();
-  },
-  watch: {
-    data() {
-      this.parseContent();
-    },
-  },
-  methods: {
-    parseContent() {
-      return `${this.data}`;
+  props: ['content'],
+  components: {VFlex},
+  computed: {
+    parsedContent() {
+      return JSON.stringify(this.content);
     },
   },
 };
